// src/app/models/user_event.interface.ts
export interface Apply {
    email: any
    user_id: number
    event_id: number
    status_id: number
    status: string
    friday_lunch: boolean
    friday_dinner: boolean
    saturday_breakfast: boolean
    saturday_lunch: boolean
    saturday_dinner: boolean
    sunday_breakfast: boolean
    sunday_lunch: boolean
    friday_room: boolean
    saturday_room: boolean
  }